swire-agent-core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3>=1.34.72
langchain==0.1.0
langchain-aws==0.1.0
pydantic==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
faiss-cpu==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
requests==2.31.0
sentence-transformers==2.2.2
pdf2image==1.16.3
Pillow==10.1.0